extract NetworkResponse to its own file
